+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.7.0
+Bugfix in DocumentDetector iOS code
+
+# 0.6.0
+Bugfix in DocumentDetector Java code
 # 0.5.0
 Bugfix in DocumentDetector Java code
 Documentation update